airflow_project/requirements.txt
apache-airflow==2.5.0
pandas
requests
beautifulsoup4
